Debugging huge load times
No luck so far. I've come to a conclusion that, irregardless of whether I solve this, tabs need to be created dynamically (together with saving)
editor-exclusive state like, say, recovery multiplier. That should also help with window creation time.

Also move Profiler to its own file.

+++ PGenEditDLL/sources/dllApi.cpp
@@ -67,7 +67,7 @@ void updatePartySizeAndPlayerPtrs()
 	{
 		players = new void* [MAX_PLAYERS];
 	}
-    assert(ptrs.size() <= MAX_PLAYERS);
+    assert((int)ptrs.size() <= MAX_PLAYERS);
     for (size_t i = 0; i < ptrs.size(); ++i)
     {
         players[i] = ptrs[i];
